Computex/requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
Django==4.2.6
django-cors-headers==4.2.0
django-polymorphic==3.1.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
docopt==0.6.2
idna==3.6
pipreqs==0.4.13
PyJWT==2.8.0
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
tzdata==2023.3
urllib3==2.1.0
yarg==0.1.9
gunicorn==20.1.0
psycopg2-binary